requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.30
pydantic>=2.9.0
python-multipart==0.0.6
pillow>=10.3.0
opencv-python==4.9.0.80
pyautogui==0.9.54
pyaudio==0.2.14
SpeechRecognition==3.10.1
requests==2.31.0
numpy>=1.24.0,<2.0.0
pynput==1.7.6
mss==9.0.1
python-dotenv==1.0.0
transformers>=4.40.0
torch>=2.0.0
accelerate==0.25.0
sentencepiece>=0.2.0
protobuf==4.25.1
easyocr==1.7.1
pytesseract==0.3.10
pyperclip==1.8.2


